previous versions/v01/requirements.txt
numpy
matplotlib
torch
scipy
PyYAML